refactor(formatter): extract MessageValidator component
- Centralise message validation and schema enforcement into dedicated MessageValidator class
- Extract methods: format_message, ensure_required_keys, string_message_field?, includes_timing?, normalise_timing
- Replace inline validation with single MessageValidator.new.validate delegation
