+++ lib/json_rails_logger.rb
@@ -9,6 +9,7 @@ require_relative 'json_rails_logger/json_formatter'
 require_relative 'json_rails_logger/formatting_components/request_message_composer'
 require_relative 'json_rails_logger/formatting_components/message_parser'
 require_relative 'json_rails_logger/formatting_components/request_context'
+require_relative 'json_rails_logger/formatting_components/message_validator'
 require_relative 'json_rails_logger/formatting_components/payload_builder'
 require_relative 'json_rails_logger/error'
 require_relative 'json_rails_logger/logger'
